Report error only once
fixes CORE-2891

Also increases the #tries for DEBUG and WARN levels, because the Nimbus
systems are taken more than 30s to start the simulation.
This should silence the spurious errors from startup and keep the actual
ones, but just reporting error once; warning will be still reported on
every iteration.