Use main loop directly when disabledAtStartup
This fixes a busy loop in the small loop when enabled_ == false on
startup (when disbled_at_startup param is set to true), because there's
no sleep.

By getting rid of enabled_ and using disabled_ attribute only, we can
indeed use the main loop for the case of disabled_at_startup == true,
and still print a ROS_WARN message before the loop.